refactored utils and updated readme
