+++ tracker/cli/balance.py
@@ -1,7 +1,7 @@
 import click
 
 from tracker import db
-from tracker.cli.utils import _template_to_title, TrackerClickGroup
+from tracker.cli.utils import template_to_title, TrackerClickGroup
 
 
 @click.group(name="balance", cls=TrackerClickGroup)
@@ -19,7 +19,7 @@ def balance_list_cmd():
     coin_balances = db.balance.get_last_coin_balances()
 
     template = "{DATE:<16}" "{COIN:<10}" "{AMOUNT:<20}"
-    click.echo("\n" + _template_to_title(template))
+    click.echo("\n" + template_to_title(template))
     for cb in coin_balances:
         click.echo(
             template.format(
